Course-02/lesson-6/requirements.txt
# Core dependencies for Refinery Optimization System
python-dotenv>=1.0.0
openai>=1.0.0
